Refactored the UserState
