+++ lib/src/features/authentication/presentation/pages/auth_page.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import '../../../../../state_management/user_state.dart';
+import '../provider/user_state.dart';
 
 class AuthPage extends StatelessWidget {
   const AuthPage({super.key});
